Add generic and Tasmota specific MQTT actors
